Use JSON for storage instead of SQL

+++ src/models/educator_model.rs
@@ -28,6 +28,6 @@ pub struct EducatorDay {
 #[serde(rename_all = "PascalCase")]
 pub struct EducatorEvents {
     pub educator_long_display_text: String,
-    pub educator_master_id: i64,
+    pub educator_master_id: u32,
     pub educator_events_days: Vec<EducatorDay>,
 }
